Modernize the architecture

+++ src/guess_buffer/create_guess_buffer.c
@@ -1,7 +1,7 @@
 #include "../../includes/guess_buffer.h"
 
 char * create_guess_buffer(int length){
-	char * buffer = (char *) malloc(sizeof(char) * (length+1));
+	char * buffer = (char *) malloc(sizeof(char) * (unsigned long)(length+1));
 	int position;
 	for (position = 0; position < length; ++position){
 		buffer[position] = PLACEHOLDER_CHARACTER;
